Chore: Migrate lib dropbox-login-shared to TypeScript
Type the DropboxApi surface that Shared touches (loginUrl/setAuthToken/
execAuthToken). The host component parameter stays `any` because
React.Component#setState uses a generic Pick<S, K> signature that's awkward
to model at this seam — desktop and mobile each have their own Props shape,
and a fragile shim would break with the next React-types bump. Switch the
two consumers (DropboxLoginScreen.tsx desktop, dropbox-login.tsx mobile)
from `require()` to a typed default import; drop the two no-explicit-any
disables on `private shared_: any`. Three pre-existing fire-and-forget
promise calls (reg.scheduleSync, two UNSAFE_componentWillMount calls of
refreshUrl) now need `void` since the wrapping methods are typed async.
