+++ packages/app-mobile/components/screens/dropbox-login.tsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { ScreenHeader } from '../ScreenHeader';
 import { _ } from '@joplin/lib/locale';
 import { BaseScreenComponent } from '../base-screen';
-const Shared = require('@joplin/lib/components/shared/dropbox-login-shared');
+import Shared from '@joplin/lib/components/shared/dropbox-login-shared';
 import shim, { MessageBoxType } from '@joplin/lib/shim';
 import { themeStyle } from '../global-style';
 
@@ -23,8 +23,7 @@ interface State {
 class DropboxLoginScreenComponent extends BaseScreenComponent<Props, State> {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Migrated from JS; per-theme StyleSheet cache
 	private styles_: Record<number, any> = {};
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Shared comes from an untyped JS module
-	private shared_: any;
+	private shared_: Shared;
 
 	public constructor(props: Props) {
 		super(props);
@@ -37,7 +36,7 @@ class DropboxLoginScreenComponent extends BaseScreenComponent<Props, State> {
 	}
 
 	public UNSAFE_componentWillMount() {
-		this.shared_.refreshUrl();
+		void this.shared_.refreshUrl();
 	}
 
 	private styles() {
